Modified ChronoTimer so that the channels array is not public. To do this each Channel has a reference to the ChronoTimer that it belongs to and it knows its own id.
Currently a channel can be triggered from the ChronoTimer or the Channel
directly. This will be good when we have an abstract API for the system
so the interface using our API can choice to use either method that
results in the same logic.

To conserve memory Channels are created initially but never
destroyed/re-created. The default state of a channel is no sensor and
disabled.